% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Install.R
\name{install_engine}
\alias{install_engine}
\title{Install Engine}
\usage{
install_engine()
}
\value{

}
\description{
This function takes no arguments, and needs to be run once in order to install the Python module
}
\examples{
install_engine()
}
