% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectPages.R
\name{selectPages}
\alias{selectPages}
\title{Select Pages}
\usage{
selectPages(pdf_file, docx_filename, pages)
}
\arguments{
\item{pdf_file}{The pdf file you wish to convert into a docx file}

\item{docx_filename}{What you wish to name the converted docx file}

\item{pages}{Here, you can pass a vector of selected pages you want to convert}
}
\value{
Converted docx file stored in the working directory
}
\description{
This function allows you to select specific pages in the pdf file you wish to convert into a docx file
}
\examples{
selectPages(pdf_file = "doc.pdf",
            docx_filename = "selected.docx",
            pages = c(10,15,20))
}
