% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startANDend.R
\name{startANDend}
\alias{startANDend}
\title{Start and End}
\usage{
startANDend(pdf_file, docx_filename, start, end)
}
\arguments{
\item{pdf_file}{The pdf file you wish to convert into a docx file}

\item{docx_filename}{What you wish to name the converted docx file}

\item{start}{The pdf file page number from which you want to start the conversion}

\item{end}{The pdf file page number at which you want to end the conversion}
}
\value{
Converted docx file stored in the working directory
}
\description{
This function allows you to choose from which page number you want the conversion to start and at which page number you want the conversion to end
}
\examples{
startANDend(pdf_file = "doc.pdf",
            docx_filename = pages.docx,
            start = 10,
            end = 20)
}
